+++ src/helpers/index.js
@@ -37,7 +37,7 @@ export { log, logger } from './log/log';
 export { sendMail } from './mail/mail';
 export { mergeSchemas, translateSchema } from './schema/schema';
 export { testRequest } from './tests/tests';
-export { getPath, getUrl, getRootUrl } from './url/url';
+export { getPath, getUrl, getUrlByPath, getRootUrl } from './url/url';
 export {
   arrayToVocabulary,
   isPromise,