+++ src/routes/actions/actions.js
@@ -5,16 +5,18 @@
 
 import { Action } from '../../models';
 
+export const handler = async (req, trx) => {
+  const actions = await Action.fetchAll({}, { order: 'order' }, trx);
+  return {
+    json: actions.toJSON(req),
+  };
+};
+
 export default [
   {
     op: 'get',
     view: '/@actions',
     permission: 'View',
-    handler: async (req, trx) => {
-      const actions = await Action.fetchAll({}, { order: 'order' }, trx);
-      return {
-        json: actions.toJSON(req),
-      };
-    },
+    handler,
   },
 ];