+++ src/routes/navroot/navroot.test.js
@@ -1,0 +1,6 @@
+import app from '../../app';
+import { testRequest } from '../../helpers';
+
+describe('Navroot', () => {
+  it('should get the navroot', () => testRequest(app, 'navroot/get'));
+});